chore(compose): move messages to separate file
